requirements.txt
python-dotenv
langchain
langchainhub
langchain_core
langchain_community
langchain_groq
streamlit
pypdf
faiss-cpu